requirements.txt
alabaster==0.7.10
appdirs==1.4.3
argcomplete==1.9.2
Arpeggio==1.6.1
Babel==2.5.1
certifi==2017.11.5
chardet==3.0.4
cx-Freeze==5.0.2
docutils==0.14
future==0.16.0
idna==2.6
imagesize==0.7.1
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
py==1.5.2
Pygments==2.2.0
PyInstaller==3.2.1
pyserial==3.0.1
pytest==3.2.5
python-escpos==2.2.0
pytz==2017.3
pyusb==1.0.0
PyYAML==3.12
qrcode==5.3
requests==2.18.4
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
textX==1.5.1
typing==3.6.2
urllib3==1.22
